+++ firedantic/_async/model.py
@@ -13,7 +13,11 @@ from google.cloud.firestore_v1.async_query import AsyncQuery
 import firedantic.operators as op
 from firedantic import async_truncate_collection
 from firedantic.configurations import CONFIGURATIONS
-from firedantic.exceptions import CollectionNotDefined, ModelNotFoundError
+from firedantic.exceptions import (
+    CollectionNotDefined,
+    InvalidDocumentID,
+    ModelNotFoundError,
+)
 
 TAsyncBareModel = TypeVar("TAsyncBareModel", bound="AsyncBareModel")
 TAsyncBareSubModel = TypeVar("TAsyncBareSubModel", bound="AsyncBareSubModel")
@@ -54,7 +58,11 @@ class AsyncBareModel(pydantic.BaseModel, ABC):
     __document_id__: str
 
     async def save(self) -> None:
-        """Saves this model in the database."""
+        """
+        Saves this model in the database.
+
+        :raise DocumentIDError: If the document ID is not valid.
+        """
         data = self.dict(by_alias=True)
         if self.__document_id__ in data:
             del data[self.__document_id__]
@@ -64,13 +72,22 @@ class AsyncBareModel(pydantic.BaseModel, ABC):
         setattr(self, self.__document_id__, doc_ref.id)
 
     async def delete(self) -> None:
-        """Deletes this model from the database."""
+        """
+        Deletes this model from the database.
+
+        :raise DocumentIDError: If the ID is not valid.
+        """
         await self._get_doc_ref().delete()
 
     def get_document_id(self):
         """
         Get the document ID for this model instance
+
+        :raise DocumentIDError: If the ID is not valid.
         """
+        doc_id = getattr(self, self.__document_id__, None)
+        if doc_id is not None:
+            self._validate_document_id(doc_id)
         return getattr(self, self.__document_id__, None)
 
     @classmethod
@@ -156,9 +173,14 @@ class AsyncBareModel(pydantic.BaseModel, ABC):
         :raise ModelNotFoundError: Raised if no matching document is found.
         """
 
-        if not doc_id:
+        try:
+            cls._validate_document_id(doc_id)
+        except InvalidDocumentID:
             # Getting a document with doc_id set to an empty string would raise a
-            # google.api_core.exceptions.InvalidArgument exception.
+            # google.api_core.exceptions.InvalidArgument exception and a doc_id
+            # containing an uneven number of slashes would raise a
+            # ValueError("A document must have an even number of path elements") and
+            # could even load data from a sub collection instead of the desired one.
             raise ModelNotFoundError(
                 f"No '{cls.__name__}' found with {cls.__document_id__} '{doc_id}'"
             )
@@ -192,8 +214,45 @@ class AsyncBareModel(pydantic.BaseModel, ABC):
         return _get_col_ref(cls, cls.__collection__)
 
     def _get_doc_ref(self) -> AsyncDocumentReference:
-        """Returns the document reference."""
+        """
+        Returns the document reference.
+
+        :raise DocumentIDError: If the ID is not valid.
+        """
         return self._get_col_ref().document(self.get_document_id())  # type: ignore
+
+    @staticmethod
+    def _validate_document_id(document_id: str):
+        """
+        Validates the Document ID is valid.
+
+        Based on information from https://firebase.google.com/docs/firestore/quotas#limits
+
+        :raise DocumentIDError: If the ID is not valid.
+        """
+        if len(document_id.encode("utf-8")) > 1500:
+            raise InvalidDocumentID("Document ID must be no longer than 1,500 bytes")
+
+        if "/" in document_id:
+            raise InvalidDocumentID("Document ID cannot contain a forward slash (/)")
+
+        if (
+            document_id.startswith("__")
+            and document_id.endswith("__")
+            and len(document_id) >= 4
+        ):
+            raise InvalidDocumentID(
+                "Document ID cannot match the regular expression __.*__"
+            )
+
+        if document_id in (".", ".."):
+            raise InvalidDocumentID(
+                "Document ID cannot solely consist of a single period (.) or double "
+                "periods (..)"
+            )
+
+        if document_id == "":
+            raise InvalidDocumentID("Document ID cannot be an empty string")
 
 
 class AsyncModel(AsyncBareModel):